♻️ Refactor: Update loading spinners
