+++ client/src/features/playlist/components/Playlist.tsx
@@ -3,9 +3,9 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../../redux/hooks.ts';
 import { useEffect } from 'react';
 import { getPlaylist } from '../playlistThunks.ts';
-import NotFound from '../../../ui/StatusScreens/NotFound.tsx';
-import ServerError from '../../../ui/StatusScreens/ServerError.tsx';
-import LoadingScreen from '../../../ui/StatusScreens/LoadingScreen.tsx';
+import NotFound from '../../../ui/statusScreens/NotFound.tsx';
+import ServerError from '../../../ui/statusScreens/ServerError.tsx';
+import LoadingScreen from '../../../ui/statusScreens/LoadingScreen.tsx';
 import PlaylistActions from './PlaylistActions.tsx';
 import ImageHeader from '../../../ui/ImageHeader/ImageHeader.tsx';
 import styles from './Playlist.module.scss';
