--- client/src/ui/StatusScreens/LoadingScreen.tsx
@@ -1,9 +1,0 @@
-const LoadingScreen = () => {
-  return (
-    <div className="h-full flex items-center justify-center text-3xl">
-      Loading...
-    </div>
-  );
-};
-
-export default LoadingScreen;
