+++ client/src/ui/statusScreens/ServerError.tsx
